+++ Lib/Utils/CommandLineParser/Definitions/GenerateTsConfigCommand.cs
@@ -1,0 +1,8 @@
+namespace Lib.Utils.CommandLineParser.Definitions;
+
+public class GenerateTsConfigCommand : CommonInteractiveCommand
+{
+    public override string[] Words => new[] { "gen-tsconfig" };
+
+    protected override string Description => "generates tsconfig.json";
+}